+++ packages/graph-explorer/src/components/Tabular/controls/ExportControl/ExternalExportControl.tsx
@@ -14,8 +14,8 @@ import {
 import { TrayArrowIcon } from "@/components/icons";
 import { TabularInstance } from "@/components/Tabular/helpers/tableInstanceToTabularInstance";
 
-import { transformToCsv } from "./transfomerToCsv";
-import { transformToJson } from "./transfomerToJson";
+import { transformToCsv } from "./transformToCsv";
+import { transformToJson } from "./transformToJson";
 import { toCsvFileData, toJsonFileData } from "@/utils/fileData";
 
 type ExportControlProps<T extends Record<string, unknown>> = {
