+++ packages/graph-explorer/src/components/Tabular/controls/ExportControl/transformToJson.ts
